repository: add SIGHUP handler only on Unix systems
